move test to util.py
