+++ pyiron_base/generic/hdfio.py
@@ -31,35 +31,6 @@ __email__ = "janssen@mpie.de"
 __status__ = "production"
 __date__ = "Sep 1, 2017"
 
-_special_symbol_replacements = {
-    '.': 'd',
-    '-': 'm',
-    '+': 'p',
-    ',': 'c',
-    ' ': '_',
-}
-
-
-def _get_safe_job_name(name, ndigits=8, extension=None):
-    def round_(value, ndigits=ndigits):
-        if isinstance(value, float) and ndigits is not None:
-            return round(value, ndigits=ndigits)
-        return value
-    if not isinstance(name, str):
-        name_rounded = [round_(nn) for nn in name]
-        job_name = '_'.join([str(nn) for nn in name_rounded])
-    else:
-        job_name = name
-    if extension is not None and job_name.endswith(extension):
-        job_name = job_name[:job_name.rfind(extension)]
-    job_name = job_name.split('/')
-    for k, v in _special_symbol_replacements.items():
-        job_name[-1] = job_name[-1].replace(k, v)
-    job_name = '/'.join(job_name)
-    if extension is not None:
-        job_name += extension
-    return job_name
-
 
 def _is_ragged_in_1st_dim_only(value: Union[np.ndarray, list]) -> bool:
     """
@@ -128,7 +99,6 @@ class FileHDFio(HasGroups, MutableMapping):
     """
 
     def __init__(self, file_name, h5_path="/", mode="a"):
-        file_name = _get_safe_job_name(file_name, extension='.h5')
         if not os.path.isabs(file_name):
             raise ValueError("file_name must be given as absolute path name")
         self._file_name = None
@@ -1090,7 +1060,6 @@ class ProjectHDFio(FileHDFio):
     """
 
     def __init__(self, project, file_name, h5_path=None, mode=None):
-        file_name = _get_safe_job_name(file_name, extension='.h5')
         self._file_name = file_name.replace("\\", "/")
         if h5_path is None:
             h5_path = "/"
