+++ tests/generic/test_fileHDFio.py
@@ -6,7 +6,6 @@ import os
 import sys
 from io import StringIO
 import numpy as np
-from pyiron_base.generic.hdfio import FileHDFio, _get_safe_job_name
 from pyiron_base.generic.hdfio import FileHDFio, _is_ragged_in_1st_dim_only
 from pyiron_base._tests import PyironTestCase
 import unittest
@@ -375,18 +374,6 @@ class TestFileHDFio(PyironTestCase):
         self.assertFalse(grp in self.i_o_hdf5.list_nodes())
         # This should not raise an error, albeit the group of hdf is removed
         hdf.remove_group()
-
-    def test_get_safe_job_name(self):
-        T = 300.8000000000002
-        p = 0.30000000000000004
-        with self.subTest("Round floats"):
-            self.assertEqual(_get_safe_job_name(('job', T, 'K', p, 'GPa')), 'job_300d8_K_0d3_GPa')
-        with self.subTest("File name with extention"):    
-            self.assertEqual(_get_safe_job_name('file_name.h5', extension='.h5'), 'file_name.h5')
-        with self.subTest("File name with extention"):
-            self.assertEqual(_get_safe_job_name('file_name', extension='.h5'), 'file_name.h5')
-        with self.subTest("unsafe string"):
-            self.assertEqual(_get_safe_job_name("2-be the.most+awful,string"), "2mbe_thedmostpawfulcstring")
 
     def test_ragged_array(self):
         """Should correctly identify ragged arrays/lists."""
