+++ tests/generic/test_units.py
@@ -86,3 +86,15 @@ class TestUnits(PyironTestCase):
         # Raise error if quantity not defined in base class
         code_units.add_quantity(quantity="force", unit=pint_registry.N)
         self.assertRaises(ValueError, UnitConverter, base_registry=base_units, code_registry=code_units)
+
+    def test_get_safe_job_name(self):
+        T = 300.8000000000002
+        p = 0.30000000000000004
+        with self.subTest("Round floats"):
+            self.assertEqual(_get_safe_job_name(('job', T, 'K', p, 'GPa')), 'job_300d8_K_0d3_GPa')
+        with self.subTest("File name with extention"):    
+            self.assertEqual(_get_safe_job_name('file_name.h5', extension='.h5'), 'file_name.h5')
+        with self.subTest("File name with extention"):
+            self.assertEqual(_get_safe_job_name('file_name', extension='.h5'), 'file_name.h5')
+        with self.subTest("unsafe string"):
+            self.assertEqual(_get_safe_job_name("2-be the.most+awful,string"), "2mbe_thedmostpawfulcstring")
